workers/requirements.txt
pika==1.3.2
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
langchain==0.0.335
openai==1.2.3
pylint==3.0.2
mypy==1.7.0
pytest==7.4.3 